testes_oficiais/gerenciador_financeiro/backend-python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
bcrypt==4.0.1
python-jose==3.3.0
python-dotenv==1.0.0
motor==3.3.2